+++ configserver/src/main/java/com/yahoo/vespa/config/server/ReloadListener.java
@@ -3,6 +3,7 @@ package com.yahoo.vespa.config.server;
 
 import com.yahoo.config.provision.ApplicationId;
 import com.yahoo.config.provision.TenantName;
+import com.yahoo.vespa.config.server.application.ApplicationSet;
 
 import java.util.Collection;
