+++ configserver/src/main/java/com/yahoo/vespa/config/server/application/ApplicationMapper.java
@@ -1,5 +1,5 @@
 // Copyright 2016 Yahoo Inc. Licensed under the terms of the Apache 2.0 license. See LICENSE in the project root.
-package com.yahoo.vespa.config.server;
+package com.yahoo.vespa.config.server.application;
 
 import com.yahoo.config.provision.ApplicationId;
 import com.yahoo.config.provision.Version;
@@ -8,7 +8,7 @@ import java.util.Map;
 import java.util.Optional;
 import java.util.concurrent.ConcurrentHashMap;
 
-import com.yahoo.vespa.config.server.application.Application;
+import com.yahoo.vespa.config.server.VersionDoesNotExistException;
 import com.yahoo.vespa.config.server.http.NotFoundException;
 
 /**
