+++ configserver/src/main/java/com/yahoo/vespa/config/server/session/RemoteSession.java
@@ -4,6 +4,7 @@ package com.yahoo.vespa.config.server.session;
 import com.yahoo.config.provision.*;
 import com.yahoo.vespa.config.server.*;
 import com.yahoo.log.LogLevel;
+import com.yahoo.vespa.config.server.application.ApplicationSet;
 import com.yahoo.vespa.config.server.modelfactory.ActivatedModelsBuilder;
 import com.yahoo.vespa.config.server.tenant.Tenants;
 import com.yahoo.vespa.curator.Curator;
