+++ configserver/src/test/java/com/yahoo/vespa/config/server/MockReloadHandler.java
@@ -2,6 +2,7 @@
 package com.yahoo.vespa.config.server;
 
 import com.yahoo.config.provision.ApplicationId;
+import com.yahoo.vespa.config.server.application.ApplicationSet;
 
 /**
  * @author lulf
