+++ configserver/src/test/java/com/yahoo/vespa/config/server/http/SessionHandlerTest.java
@@ -11,7 +11,7 @@ import com.yahoo.log.LogLevel;
 import com.yahoo.path.Path;
 import com.yahoo.container.jdisc.HttpRequest;
 import com.yahoo.container.jdisc.HttpResponse;
-import com.yahoo.vespa.config.server.ApplicationSet;
+import com.yahoo.vespa.config.server.application.ApplicationSet;
 import com.yahoo.vespa.config.server.HostRegistry;
 import com.yahoo.config.provision.ApplicationId;
 import com.yahoo.config.provision.TenantName;
